increase mem and put a little error logic

+++ scripts/workers/dedup.sh
@@ -2,7 +2,7 @@
 
 #PBS -W group_list=bhurwitz
 #PBS -q qualified
-#PBS -l select=1:ncpus=6:mem=36gb
+#PBS -l select=1:ncpus=12:mem=72gb
 ###and the amount of time required to run it
 #PBS -l walltime=24:00:00
 #PBS -l cput=24:00:00
@@ -32,14 +32,19 @@ cd $GROUPED_DIR
 
 echo Doing sample $SAMPLE
 
-set -x
+#set -x
 
-echo Concatenating all fastqs into temp file
-time cat "$SAMPLE".1.fastq "$SAMPLE".2.fastq "$SAMPLE".nomatch.fastq > "$SAMPLE".temp
+#echo Concatenating all fastqs into temp file
+#time cat "$SAMPLE".1.fastq "$SAMPLE".2.fastq "$SAMPLE".nomatch.fastq > "$SAMPLE".temp
 echo Filtering out duplicate reads
 time fastq filter --adjust 64 --unique "$SAMPLE".temp \
     > $DEDUPED_DIR/"$SAMPLE".deduped.fastq
-echo Removing temp file
-rm "$SAMPLE".temp
+if [[ $? -eq 0 ]]; then
+    echo Removing temp file
+    rm "$SAMPLE".temp
+else
+    echo Something went wrong with fastq filter
+    echo Keeping temp file for another try
+fi
 
 echo Done at $(date)
